SXTgetdata

SXTgetdata is used to split data into tags, subject and QC.

SXTgetdata <- function(filename = "Metabolomics data", polarity = "none",
                          minfrac.qc = 0.8, minfrac.sample = 0.5, filter = "no",
                          normalization.method = "svr",
                          optimization = TRUE, begin = 0.5, end = 1,step = 0.2,
                          ##loess parameters
                          multiple = 5,
                          ##svr parameters
                          rerun.loess = TRUE, rerun.svr = TRUE,
                          peakplot=TRUE, datastyle = "tof", dimension1 = TRUE, user="other")

Details

Package:	SXTgetdata

Type:	Package

Version:	1.1

Date:	2016-04-27

License:	Nothing

Depends:	R (> = 3.1.1)

Imports:	e1071

Version 1.1:	(1) The data after normalization can be given dimension. See the Argument: dimensition1. (2) The injection order can be got from a new file named as "worklist.csv". See the example. (3) It can process positive and negative data together. See the Argument: polarity.

Value

peak plot:	A folder contains all the peaks plot before and sfter SVR normalization.

data svr nor.csv:	A csv data after SVR normalization

normalization file:	The intermediate data

RSD compare plot.jpeg:	A figure show the comparing of RSDs of peaks before and after SVR normalization

RSD distribution.csv:	RSD distributions of data before and after SVR normalization

RSD distribution.jpeg:	A figure show the RSD distributions before and after SVR normalization

rsd.change.csv:	A table show the RSDs of peaks before and after SVR normalization

References

Normalization and Integration of Large-Scale Metabolomics Data Using Support Vector Regression

See also

Examples